# BUG: Fixes GH9311 groupby on datetime64