datetime64 columns were changing at the nano-second scale when
applying a groupby aggregator.

closes #9311 
closes #6620