@shoyer 2015-01-25
This looks like some nice fixes but also looks likely under-tested -- you've changed a lot of functions and only added tests for one of them.