@shoyer 2015-01-25
I would probably scale this back to only change operations without any possibility of precision/overflow issues -- I think that's `first`/`last`/`min`/`max`/`nth`. The arithmetic `mean`/`sum`/`var`/`std` should probably wait until we define it even on datetime64 without grouping, as @jreback mentions above (though efforts there would certainly be appreciated).

Also, it is almost impossible to write too many tests :).