@iwschris 2015-01-26
Awesome. Thanks for the feedback. I'll have updates, and more tests in here shortly.