@shoyer 2015-01-30
@iwschris iNaT -> NaT happens automatically when you cast back to datetime64. You can use `values.view('datetime64[ns]')` for that, assuming values is an array with int64 dtype. This won't work for integers -- you'll need to upcast those to float.