@jreback 2015-01-30
@iwschris just a do the view as @shoyer suggest
you know you have an int64 (or int32) so this will work
(only take a view if it's actually a datetime64/tiemselta64) in the first place
this should be done in python for consistency (we just normally return basic types to/from cython)