@iwschris 2015-02-02
@shoyer -- Pretty sure this is ready for review again. I was able to keep the safe casting turned on in internals.py and I validated that vbench is still looking pretty good.

Here's the output from vbench:

```

Invoked with :
--ncalls: 3
--repeats: 3


-------------------------------------------------------------------------------
Test name                                    | head[ms] | base[ms] |  ratio   |
-------------------------------------------------------------------------------
index_from_series_ctor                       |   0.0147 |   0.0337 |   0.4363 |
series_constructor_ndarray                   |   0.0137 |   0.0277 |   0.4943 |
concat_empty_frames1                         |   0.7093 |   1.3227 |   0.5363 |
frame_get_numeric_data                       |   0.0956 |   0.1766 |   0.5414 |
join_dataframe_index_single_key_small        |   7.9147 |  13.9790 |   0.5662 |
groupby_ngroups_100_first                    |   0.2720 |   0.4776 |   0.5694 |
read_csv_infer_datetime_format_custom        |   7.8737 |  11.8280 |   0.6657 |
join_dataframe_index_single_key_bigger       |   8.9413 |  12.7583 |   0.7008 |
groupby_ngroups_100_last                     |   0.2457 |   0.3486 |   0.7046 |
groupby_ngroups_10000_median                 |   2.1223 |   2.9956 |   0.7085 |
frame_fancy_lookup_all                       |  10.6823 |  14.9640 |   0.7139 |
groupby_ngroups_100_max                      |   0.2611 |   0.3640 |   0.7172 |
groupby_ngroups_100_min                      |   0.2487 |   0.3421 |   0.7270 |
groupby_ngroups_100_count                    |   0.2720 |   0.3714 |   0.7325 |
frame_from_series                            |   0.1073 |   0.1454 |   0.7381 |
ctor_index_array_string                      |   0.0143 |   0.0193 |   0.7407 |
dataframe_resample_mean_numpy                |   2.4650 |   3.2423 |   0.7603 |
frame_iteritems_cached                       |   0.4021 |   0.5227 |   0.7692 |
index_float64_boolean_indexer                |   2.8100 |   3.6487 |   0.7701 |
frame_constructor_ndarray                    |   0.0670 |   0.0866 |   0.7734 |
reindex_frame_level_align                    |   0.5527 |   0.6983 |   0.7914 |
frame_count_level_axis0_multi                |  53.3690 |  66.9137 |   0.7976 |
timeseries_custom_bmonthbegin_incr_n         |   0.1923 |   0.2387 |   0.8059 |
frame_ctor_dtindex_BYearEndx1                |   1.1930 |   1.4800 |   0.8061 |
frame_reindex_axis0                          | 123.4897 | 152.9363 |   0.8075 |
frame_assign_timeseries_index                |   0.5610 |   0.6904 |   0.8126 |
read_csv_infer_datetime_format_iso8601       |   1.4773 |   1.8067 |   0.8177 |
groupby_frame_cython_many_columns            |   1.9461 |   2.3790 |   0.8180 |
frame_repr_wide                              |  10.0773 |  12.3091 |   0.8187 |
frame_ctor_dtindex_BDayx2                    |   0.9710 |   1.1800 |   0.8229 |
frame_nonunique_equal                        |   6.3940 |   7.7110 |   0.8292 |
frame_ctor_dtindex_QuarterBeginx1            |   0.9793 |   1.1793 |   0.8304 |
reindex_fillna_pad                           |   0.2143 |   0.2580 |   0.8306 |
frame_apply_ref_by_name                      |  10.7396 |  12.9287 |   0.8307 |
groupby_ngroups_10000_max                    |   1.5074 |   1.8050 |   0.8351 |
groupby_ngroups_10000_first                  |   1.5857 |   1.8973 |   0.8358 |
groupby_ngroups_100_cummax                   |  11.6933 |  13.9883 |   0.8359 |
index_datetime_intersection                  |   8.0390 |   9.6114 |   0.8364 |
frame_mask_floats                            |   4.5877 |   5.4599 |   0.8403 |
frame_drop_duplicates                        |  11.3926 |  13.4079 |   0.8497 |
groupby_ngroups_100_std                      |   0.3264 |   0.3827 |   0.8530 |
frame_reindex_columns                        |   0.2379 |   0.2789 |   0.8530 |
groupby_multi_size                           |  17.3423 |  20.3076 |   0.8540 |
series_ix_slice                              |   0.0416 |   0.0486 |   0.8562 |
datetimeindex_unique                         |   0.0723 |   0.0843 |   0.8577 |
groupby_first_float64                        |   2.2407 |   2.6090 |   0.8588 |
frame_ctor_dtindex_BQuarterEndx2             |   1.1080 |   1.2867 |   0.8611 |
reindex_frame_level_reindex                  |   0.5770 |   0.6690 |   0.8624 |
timeseries_custom_bmonthbegin_decr_n         |   0.2031 |   0.2350 |   0.8641 |
frame_ctor_dtindex_MonthEndx1                |   1.0313 |   1.1810 |   0.8732 |
groupby_series_nth_none                      |   1.0096 |   1.1547 |   0.8744 |
groupby_ngroups_100_rank                     |  11.8863 |  13.5717 |   0.8758 |
stat_ops_frame_sum_float_axis_1              |   3.5137 |   3.9994 |   0.8785 |
series_ctor_from_dict                        |   2.0247 |   2.3017 |   0.8796 |
frame_ctor_dtindex_MonthEndx2                |   1.0126 |   1.1484 |   0.8818 |
groupby_frame_singlekey_integer              |   1.3860 |   1.5713 |   0.8821 |
frame_ctor_dtindex_BYearBeginx1              |   1.1864 |   1.3447 |   0.8823 |
frame_drop_dup_na_inplace                    |   1.7404 |   1.9697 |   0.8836 |
groupby_frame_nth_any                        |   4.4223 |   4.9994 |   0.8846 |
frame_drop_dup_inplace                       |   1.8404 |   2.0734 |   0.8876 |
reindex_daterange_pad                        |   0.5534 |   0.6224 |   0.8892 |
groupby_ngroups_10000_var                    |   1.6543 |   1.8600 |   0.8894 |
timeseries_infer_freq                        |   6.6311 |   7.4290 |   0.8926 |
frame_interpolate_some_good                  |   0.9607 |   1.0746 |   0.8940 |
groupby_ngroups_100_cummin                   |  10.7533 |  12.0024 |   0.8959 |
read_csv_standard                            |   8.7403 |   9.7334 |   0.8980 |
timeseries_to_datetime_iso8601               |   3.1000 |   3.4483 |   0.8990 |
timeseries_to_datetime_YYYYMMDD              |   6.3020 |   6.9950 |   0.9009 |
sql_string_write_sqlalchemy                  |  54.9703 |  60.7810 |   0.9044 |
groupby_ngroups_10000_count                  |   1.5697 |   1.7330 |   0.9058 |
groupby_ngroups_10000_diff                   | 905.3814 | 999.3493 |   0.9060 |
frame_html_repr_trunc_mi                     |  29.4023 |  32.4394 |   0.9064 |
groupby_ngroups_10000_std                    |   1.8214 |   2.0063 |   0.9078 |
groupby_ngroups_10000_sem                    |   2.1150 |   2.3276 |   0.9087 |
frame_from_records_generator_nrows           |   0.6553 |   0.7210 |   0.9090 |
index_int64_union                            |  52.9044 |  58.1080 |   0.9104 |
multiindex_from_product                      |   8.3253 |   9.1114 |   0.9137 |
index_datetime_union                         |   7.9623 |   8.7016 |   0.9150 |
frame_ctor_dtindex_Weekx2                    |   0.7860 |   0.8587 |   0.9153 |
groupby_nth_object_any                       | 872.7303 | 951.7753 |   0.9169 |
groupby_ngroups_100_mean                     |   0.2930 |   0.3186 |   0.9197 |
packers_read_pack                            |  62.9237 |  68.2081 |   0.9225 |
groupby_frame_apply_overhead                 |   5.9593 |   6.4523 |   0.9236 |
timeseries_large_lookup_value                |   0.0126 |   0.0137 |   0.9244 |
packers_write_json_mixed_float_int_str       |  93.9660 | 101.5189 |   0.9256 |
frame_ctor_dtindex_BYearBeginx2              |   1.1083 |   1.1961 |   0.9266 |
eval_frame_add_python_one_thread             |  15.0924 |  16.2520 |   0.9286 |
frame_iloc_dups                              |   0.1764 |   0.1896 |   0.9300 |
groupby_ngroups_10000_sum                    |   1.8277 |   1.9616 |   0.9317 |
groupby_ngroups_10000_last                   |   1.7217 |   1.8463 |   0.9325 |
packers_read_stata                           |  36.7910 |  39.4524 |   0.9325 |
read_csv_precise_converter                   |   1.1957 |   1.2817 |   0.9328 |
timeseries_asof                              |   5.7844 |   6.1994 |   0.9331 |
series_value_counts_int64                    |   1.5830 |   1.6963 |   0.9332 |
frame_dtypes                                 |   0.0799 |   0.0857 |   0.9332 |
panel_shift_minor                            |   0.0700 |   0.0750 |   0.9333 |
groupby_dt_size                              |  18.6746 |  19.9850 |   0.9344 |
timeseries_custom_bday_cal_incr_neg_n        |   0.0196 |   0.0210 |   0.9356 |
index_str_boolean_indexer                    |   6.8731 |   7.3450 |   0.9357 |
groupby_first_float32                        |   2.4620 |   2.6309 |   0.9358 |
stat_ops_frame_sum_float_axis_0              |   3.4143 |   3.6473 |   0.9361 |
groupby_multi_different_functions            |   8.2520 |   8.8123 |   0.9364 |
timeseries_iter_datetimeindex_preexit        |   9.0630 |   9.6117 |   0.9429 |
frame_repr_tall                              |  17.1143 |  18.1334 |   0.9438 |
packers_read_json_date_index                 | 139.4784 | 147.7621 |   0.9439 |
dtype_infer_datetime64                       |   5.7657 |   6.0943 |   0.9461 |
read_csv_roundtrip_converter                 |   1.8900 |   1.9963 |   0.9468 |
groupby_ngroups_100_prod                     |   0.3593 |   0.3793 |   0.9472 |
lib_fast_zip_fillna                          |   8.9866 |   9.4833 |   0.9476 |
frame_reindex_both_axes_ix                   |  27.4524 |  28.9690 |   0.9476 |
sort_level_zero                              |   8.5460 |   9.0003 |   0.9495 |
groupby_ngroups_10000_cumprod                | 987.3126 | 1039.5793 |   0.9497 |
groupby_indices                              |   4.2933 |   4.5194 |   0.9500 |
frame_ctor_dtindex_MonthBeginx2              |   1.0253 |   1.0786 |   0.9506 |
series_getitem_slice                         |   0.0339 |   0.0357 |   0.9510 |
read_csv_skiprows                            |  11.6111 |  12.2014 |   0.9516 |
sql_float_write_fallback                     |  23.6900 |  24.8843 |   0.9520 |
sql_write_sqlalchemy                         | 121.5783 | 127.5000 |   0.9536 |
packers_read_json                            | 142.3674 | 149.2996 |   0.9536 |
sql_float_read_table_sqlalchemy              |  11.9420 |  12.5156 |   0.9542 |
frame_ctor_dtindex_Weekx1                    |   0.8196 |   0.8589 |   0.9542 |
groupby_ngroups_100_head                     |   0.5653 |   0.5923 |   0.9544 |
join_non_unique_equal                        |   0.3424 |   0.3587 |   0.9546 |
groupby_multi_different_numpy_functions      |   8.0523 |   8.4341 |   0.9547 |
merge_2intkey_sort                           |  24.8954 |  26.0690 |   0.9550 |
frame_ctor_dtindex_BMonthBeginx1             |   1.1203 |   1.1717 |   0.9562 |
groupby_transform_multi_key2                 |  33.9526 |  35.4800 |   0.9570 |
series_timestamp_compare                     |   7.5290 |   7.8657 |   0.9572 |
packers_write_stata_with_validation          |  31.8860 |  33.2757 |   0.9582 |
frame_ctor_dtindex_DateOffsetx2              |   0.8277 |   0.8633 |   0.9588 |
strings_count                                |   4.7774 |   4.9780 |   0.9597 |
sql_float_read_query_fallback                |   6.5816 |   6.8550 |   0.9601 |
frame_apply_pass_thru                        |   3.2713 |   3.4067 |   0.9603 |
index_float64_div                            |   2.0644 |   2.1497 |   0.9603 |
frame_multi_and_no_ne                        |  20.8013 |  21.6600 |   0.9604 |
index_int64_intersection                     |  10.9847 |  11.4144 |   0.9624 |
frame_ctor_dtindex_DateOffsetx1              |   0.8663 |   0.8997 |   0.9629 |
timeseries_iter_periodindex_preexit          |   9.4426 |   9.8027 |   0.9633 |
strings_len                                  |   1.3343 |   1.3837 |   0.9643 |
frame_to_csv                                 | 110.6810 | 114.6793 |   0.9651 |
strings_cat                                  |   0.5057 |   0.5236 |   0.9657 |
frame_count_level_axis0_mixed_dtypes_multi   | 107.9280 | 111.7450 |   0.9658 |
join_dataframe_index_multi                   |  12.2537 |  12.6839 |   0.9661 |
groupby_ngroups_10000_mad                    | 3435.5440 | 3555.5203 |   0.9663 |
sql_float_write_sqlalchemy                   |  55.9303 |  57.8830 |   0.9663 |
frame_ctor_dtindex_Easterx1                  |   0.9393 |   0.9716 |   0.9667 |
groupby_ngroups_100_any                      |   6.9167 |   7.1410 |   0.9686 |
stats_rank2d_axis1_average                   |   8.8743 |   9.1600 |   0.9688 |
groupby_ngroups_10000_describe               | 12356.2800 | 12750.7817 |   0.9691 |
groupby_multi_series_op                      |  10.1536 |  10.4600 |   0.9707 |
packers_write_json                           |  73.1703 |  75.3726 |   0.9708 |
series_getitem_array                         |   0.3147 |   0.3237 |   0.9723 |
frame_dropna_axis1_any_mixed_dtypes          | 164.5314 | 169.1933 |   0.9724 |
frame_dropna_axis0_any_mixed_dtypes          | 159.9306 | 164.4520 |   0.9725 |
concat_small_frames                          |  40.3993 |  41.5320 |   0.9727 |
timeseries_timestamp_downsample_mean         |   3.4340 |   3.5297 |   0.9729 |
frame_ctor_dtindex_BusinessDayx2             |   0.9197 |   0.9443 |   0.9739 |
groupby_transform_series2                    | 108.0357 | 110.8853 |   0.9743 |
groupby_ngroups_100_sum                      |   0.3707 |   0.3804 |   0.9745 |
frame_ctor_dtindex_YearEndx1                 |   0.8994 |   0.9213 |   0.9762 |
series_xs_mi_ix                              |   2.9663 |   3.0383 |   0.9763 |
packers_write_csv                            | 943.0904 | 965.2163 |   0.9771 |
timeseries_asof_single                       |   0.0173 |   0.0177 |   0.9776 |
frame_dropna_axis0_all_mixed_dtypes          | 192.6063 | 196.9613 |   0.9779 |
groupby_nth_datetimes_any                    | 889.3877 | 909.3717 |   0.9780 |
frame_ctor_dtindex_CDayx2                    |   0.8940 |   0.9136 |   0.9785 |
frame_interpolate                            |  64.3644 |  65.7176 |   0.9794 |
timeseries_iter_datetimeindex                | 539.4704 | 550.7627 |   0.9795 |
frame_to_string_floats                       |  22.1376 |  22.5883 |   0.9800 |
eval_frame_mult_python                       |  16.0577 |  16.3810 |   0.9803 |
frame_ctor_dtindex_CBMonthBeginx2            |   2.0204 |   2.0594 |   0.9811 |
panel_shift                                  |   0.0716 |   0.0730 |   0.9815 |
frame_ctor_dtindex_BQuarterEndx1             |   1.0660 |   1.0850 |   0.9826 |
reindex_daterange_backfill                   |   0.7203 |   0.7331 |   0.9827 |
packers_write_json_date_index                |  83.0213 |  84.4580 |   0.9830 |
timeseries_custom_bmonthend_decr_n           |   0.2303 |   0.2340 |   0.9840 |
frame_mask_bools                             |   6.4433 |   6.5400 |   0.9852 |
timeseries_add_irregular                     |   9.7550 |   9.8964 |   0.9857 |
frame_apply_lambda_mean                      |   4.3870 |   4.4500 |   0.9858 |
series_drop_duplicates_string                |   0.4007 |   0.4063 |   0.9861 |
frame_ctor_nested_dict                       |  56.1170 |  56.9040 |   0.9862 |
frame_ctor_dtindex_YearBeginx1               |   0.8856 |   0.8980 |   0.9863 |
groupby_pivot_table                          |  12.7386 |  12.9116 |   0.9866 |
frame_interpolate_some_good_infer            |   2.0733 |   2.1013 |   0.9866 |
frame_ctor_dtindex_YearEndx2                 |   0.9444 |   0.9566 |   0.9872 |
groupby_ngroups_100_all                      |   7.2843 |   7.3783 |   0.9873 |
groupby_ngroups_10000_cumsum                 | 987.8446 | 1000.3951 |   0.9875 |
groupby_ngroups_100_sem                      |   0.6073 |   0.6146 |   0.9881 |
groupby_transform_multi_key1                 |  49.5160 |  50.1057 |   0.9882 |
melt_dataframe                               |   1.4334 |   1.4487 |   0.9894 |
stat_ops_frame_sum_int_axis_1                |   3.4297 |   3.4643 |   0.9900 |
strings_contains_many                        |   4.6546 |   4.6970 |   0.9910 |
frame_mult                                   |   4.3540 |   4.3923 |   0.9913 |
groupby_frame_apply                          |  29.8660 |  30.1157 |   0.9917 |
panel_pct_change_major                       | 5035.1934 | 5074.3450 |   0.9923 |
read_csv_comment2                            |  20.2267 |  20.3773 |   0.9926 |
datetimeindex_add_offset                     |   0.1920 |   0.1934 |   0.9930 |
strings_endswith                             |   3.2713 |   3.2940 |   0.9931 |
frame_apply_np_mean                          |   4.2990 |   4.3247 |   0.9941 |
datetimeindex_normalize                      |   2.6603 |   2.6757 |   0.9942 |
frame_dropna_axis1_any                       |  19.4923 |  19.6040 |   0.9943 |
packers_write_json_mixed_delta_int_tstamp    | 102.4027 | 102.9747 |   0.9944 |
read_csv_vb                                  |  16.3047 |  16.3777 |   0.9955 |
groupby_ngroups_10000_cumcount               |  57.6493 |  57.8650 |   0.9963 |
frame_object_equal                           |   6.4127 |   6.4360 |   0.9964 |
frame_ctor_dtindex_Easterx2                  |   0.9423 |   0.9456 |   0.9965 |
series_ix_scalar                             |   0.0236 |   0.0237 |   0.9966 |
write_csv_standard                           |  38.6047 |  38.7180 |   0.9971 |
groupby_transform_multi_key4                 | 100.7117 | 100.9640 |   0.9975 |
groupby_ngroups_10000_size                   |   3.2973 |   3.3036 |   0.9981 |
groupby_nth_float32_none                     |  66.8260 |  66.9477 |   0.9982 |
timeseries_with_format_no_exact              | 601.0760 | 601.8333 |   0.9987 |
frame_fillna_many_columns_pad                |   3.2874 |   3.2890 |   0.9995 |
datetimeindex_infer_dst                      |   2.6283 |   2.6280 |   1.0001 |
panel_from_dict_same_index                   |  30.4280 |  30.4210 |   1.0002 |
groupby_ngroups_10000_value_counts           | 3861.8680 | 3859.5047 |   1.0006 |
timeseries_asof_nan                          |   5.5869 |   5.5803 |   1.0012 |
frame_multi_and                              |  22.1247 |  22.0914 |   1.0015 |
frame_add_st                                 |   4.3914 |   4.3824 |   1.0020 |
groupby_frame_median                         |   4.9524 |   4.9326 |   1.0040 |
frame_fancy_lookup                           |   2.5950 |   2.5843 |   1.0042 |
frame_ctor_dtindex_QuarterEndx1              |   1.0297 |   1.0243 |   1.0052 |
groupby_last_object                          |  12.3967 |  12.3320 |   1.0052 |
timeseries_is_month_start                    |   2.3900 |   2.3774 |   1.0053 |
series_align_left_monotonic                  |  12.2104 |  12.1450 |   1.0054 |
lib_fast_zip                                 |   6.3469 |   6.3126 |   1.0054 |
frame_ctor_dtindex_CBMonthEndx2              |   2.8620 |   2.8457 |   1.0057 |
groupby_ngroups_10000_pct_change             | 2949.9527 | 2930.2514 |   1.0067 |
frame_fillna_inplace                         |   7.3990 |   7.3454 |   1.0073 |
eval_frame_add_python                        |  14.7847 |  14.6750 |   1.0075 |
groupby_nth_object_none                      | 479.3561 | 475.7130 |   1.0077 |
timeseries_slice_minutely                    |   0.0400 |   0.0397 |   1.0080 |
frame_mult_st                                |   4.3757 |   4.3403 |   1.0081 |
groupby_multi_python                         |  64.3490 |  63.8003 |   1.0086 |
frame_reindex_both_axes                      |  27.2237 |  26.9873 |   1.0088 |
dataframe_reindex                            |   0.2473 |   0.2450 |   1.0094 |
groupby_nth_datetimes_none                   | 460.5076 | 456.1773 |   1.0095 |
frame_ctor_dtindex_QuarterEndx2              |   1.0264 |   1.0166 |   1.0096 |
frame_iloc_big                               |   0.1190 |   0.1176 |   1.0115 |
sql_datetime_write_sqlalchemy                | 105.9604 | 104.7047 |   1.0120 |
…